# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fair Drop API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/me/admin": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Verify admin access",
        "description": "Returns caller profile if they hold ADMIN role in the profiles table, otherwise returns 403 Forbidden.",
        "operationId": "get_me_admin_api_v1_auth_me_admin_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/campaigns": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create campaign (DRAFT)",
        "description": "Creates a new campaign in DRAFT state. Seats are not generated until prepare step.",
        "operationId": "create_campaign_endpoint_api_v1_admin_campaigns_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all campaigns (admin)",
        "description": "Returns all campaigns including DRAFT and PREPARING, with aggregated seat and registration metrics.",
        "operationId": "list_campaigns_admin_api_v1_admin_campaigns_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignAdminResponse"
                  },
                  "title": "Response List Campaigns Admin Api V1 Admin Campaigns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get campaign detail (admin)",
        "description": "Returns full administrative view of campaign with inventory counts and pauses.",
        "operationId": "get_campaign_admin_api_v1_admin_campaigns__campaign_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update campaign (DRAFT or PREPARING only)",
        "description": "Updates policy and metadata fields while campaign is in DRAFT or PREPARING state.",
        "operationId": "update_campaign_endpoint_api_v1_admin_campaigns__campaign_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/prepare": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Prepare campaign and generate seat inventory",
        "description": "Transitions DRAFT -> PREPARING, calculates policy hash, and bulk-inserts initial seat inventory.",
        "operationId": "prepare_campaign_endpoint_api_v1_admin_campaigns__campaign_id__prepare_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/publish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Publish and open campaign",
        "description": "Transitions PREPARING -> OPEN. Makes campaign visible to the public and open for registrations.",
        "operationId": "publish_campaign_endpoint_api_v1_admin_campaigns__campaign_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/close": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Close campaign registrations",
        "description": "Transitions OPEN -> CLOSED. Halts all new registrations.",
        "operationId": "close_campaign_endpoint_api_v1_admin_campaigns__campaign_id__close_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/freeze": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Freeze campaign policy and roster",
        "description": "Transitions CLOSED -> FROZEN, captures policy snapshot, and locks eligible participant roster.",
        "operationId": "freeze_campaign_endpoint_api_v1_admin_campaigns__campaign_id__freeze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreezeRosterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/draw": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Execute deterministic uniform lottery draw",
        "description": "Executes verifiable lottery draw, creates LotteryEntries and Entitlements, and transitions FROZEN -> CLAIMING.",
        "operationId": "draw_campaign_endpoint_api_v1_admin_campaigns__campaign_id__draw_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LotteryDrawRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotteryDrawResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/pause": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Pause operational scope",
        "description": "Pauses ADMISSION, REGISTRATION, or REDEMPTION without altering lifecycle state.",
        "operationId": "pause_campaign_endpoint_api_v1_admin_campaigns__campaign_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/campaigns/{campaign_id}/resume": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Resume operational scope",
        "description": "Resumes ADMISSION, REGISTRATION, or REDEMPTION.",
        "operationId": "resume_campaign_endpoint_api_v1_admin_campaigns__campaign_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "event_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Start"
          },
          "capacity": {
            "type": "integer",
            "maximum": 100000.0,
            "exclusiveMinimum": 0.0,
            "title": "Capacity"
          },
          "registration_start": {
            "type": "string",
            "format": "date-time",
            "title": "Registration Start"
          },
          "registration_end": {
            "type": "string",
            "format": "date-time",
            "title": "Registration End"
          },
          "redemption_deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Redemption Deadline"
          },
          "max_tickets_per_participant": {
            "type": "integer",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Tickets Per Participant",
            "default": 1
          },
          "allocation_method": {
            "type": "string",
            "const": "UNIFORM_LOTTERY",
            "title": "Allocation Method",
            "default": "UNIFORM_LOTTERY"
          },
          "standby_policy": {
            "type": "string",
            "const": "FIXED_ORDER",
            "title": "Standby Policy",
            "default": "FIXED_ORDER"
          },
          "policy_version": {
            "type": "string",
            "maxLength": 50,
            "title": "Policy Version",
            "default": "v1.0"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "capacity",
          "registration_start",
          "registration_end",
          "redemption_deadline"
        ],
        "title": "CampaignCreateRequest",
        "description": "Payload for creating a new campaign in DRAFT state."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Supabase user UUID"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "User display name"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "title": "Role",
            "description": "Platform authorization role"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "description": "Email verification status"
          },
          "phone_verified": {
            "type": "boolean",
            "title": "Phone Verified",
            "description": "Phone verification status"
          },
          "institute_verified": {
            "type": "boolean",
            "title": "Institute Verified",
            "description": "Institute verification status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Account creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "email_verified",
          "phone_verified",
          "institute_verified",
          "created_at"
        ],
        "title": "ProfileResponse",
        "description": "Public profile representation."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}